+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/model/entity/Favorite.java
