+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/model/entity/Role.java
@@ -1,4 +1,4 @@
-package com.parkrangers.parkquest_backend.model;
+package com.parkrangers.parkquest_backend.model.entity;
 
 import jakarta.persistence.*;
