+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/model/entity/User.java
@@ -3,6 +3,7 @@ package com.parkrangers.parkquest_backend.model.entity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -17,24 +18,50 @@ import java.util.Collections;
 @AllArgsConstructor
 public class User implements UserDetails {
 
+    @Getter
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     private Long id;
 
+    @Getter
+    @Column(name = "googleId", nullable = true, unique = true)
+    private String googleId;
+
     @Column(name = "username", nullable = false, unique = true)
     private String username;
 
+    @Getter
     @Column(name = "email", nullable = false, unique = true)
     private String email;
 
-    @Column(name = "password", nullable = false)
+    @Column(name = "password")
     private String password;
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         // You can define roles and permissions if necessary.
         return Collections.emptyList();  // If you don't use roles, just return an empty list
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public void setGoogleId(String googleId) {
+        this.googleId = googleId;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
     }
 
     @Override
