+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/model/request/ParkReviewRequest.java
@@ -1,13 +1,13 @@
-package com.parkrangers.parkquest_backend.model.request;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ParkReviewRequest {
-
-    private String parkId;      // parkId as a simple string
-    private String reviewText;  // Review content
-    private int rating;         // Rating out of 5
-}
+//package com.parkrangers.parkquest_backend.model.request;
+//
+//import lombok.Getter;
+//import lombok.Setter;
+//
+//@Getter
+//@Setter
+//public class ParkReviewRequest {
+//
+//    private String parkId;      // parkId as a simple string
+//    private String reviewText;  // Review content
+//    private int rating;         // Rating out of 5
+//}
