+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/repository/ParkReviewRepository.java
@@ -1,10 +1,12 @@
 //package com.parkrangers.parkquest_backend.repository;
 //
-//import com.parkrangers.parkquest_backend.model.ParkReview;
+//import com.parkrangers.parkquest_backend.model.entity.ParkReview;
 //import org.springframework.data.jpa.repository.JpaRepository;
 //
 //import java.util.List;
+//import java.util.Optional;
 //
 //public interface ParkReviewRepository extends JpaRepository<ParkReview, Long> {
-//    List<ParkReview> findbyParkId(String parkId); // Find reviews by parkId
+//    List<ParkReview> findByParkId(String parkId);
+//    Optional<com.parkrangers.parkquest_backend.model.entity.ParkReview> findByReviewId(Long reviewId);
 //}
