+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/repository/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.parkrangers.parkquest_backend.repository;
 
-import com.parkrangers.parkquest_backend.model.Role;
+import com.parkrangers.parkquest_backend.model.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
