+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.parkrangers.parkquest_backend.repository;
 
-import com.parkrangers.parkquest_backend.model.User;
+import com.parkrangers.parkquest_backend.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
